requirements.txt
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
fastapi==0.62.0
h11==0.11.0
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
six==1.15.0
starlette==0.13.6
uvicorn==0.13.1
